pyproject.toml
[project]
name = "langraph-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi[all]>=0.115.8",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.7",
    "langgraph>=0.2.74",
    "pypdf>=5.3.0",
    "sseclient>=0.0.27",
    "streamlit>=1.42.2",
    "uvicorn>=0.34.0",
]
